requirements.txt
easyocr==1.6.2
numpy==1.23.5
PyQt5==5.15.7
Pillow==9.3.0
rembg==2.0.30
